+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v3.31.0 - 2023-09-14
+
+- Added a `--max-runs` flag, which sets the maximum number of times a task
+  should run before being considered an infinite loop or a cyclic dep,
+  and therefore killed (#1321 by @bogwro).
+- Added a test task in `testdata/for` and test case to test `--max-runs` behavior.
+- Fixed a bug where `TaskCalledTooManyTimes` error always shows 0 as number
+  of exceeded max runs.
+
 ## v3.30.0 - 2023-09-13
 
 - Prep work for Remote Taskfiles (#1316 by @pd93).